updated styling and restructure of buttons
